backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
openai-whisper==20231117
anthropic==0.25.0
supabase==2.4.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
langdetect==1.0.9
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9

torch
transformers
openai-whisper==20231117
fasttext

tldextract
python-whois
regex

supabase==2.4.0
python-dotenv==1.0.1
httpx==0.27.0

numpy
pydantic==2.7.1